--- JavaSelenium/src/TestSample.java
@@ -1,40 +1,0 @@
-import static org.junit.Assert.*;
-
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class TestSample {
-	
-	WebDriver driver;
-
-	@BeforeClass
-	public static void setUpBeforeClass() throws Exception {
-	}
-
-	@AfterClass
-	public static void tearDownAfterClass() throws Exception {
-	}
-
-	@Before
-	public void setUp() throws Exception {
-		System.setProperty("webdriver.chrome.driver", "C:\\Users\\jfl2\\Desktop\\chromedriver.exe");
-		driver = new ChromeDriver();
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		driver.close();
-		driver.quit();
-	}
-
-	@Test
-	public void test() {
-		driver.get("https://opensource-demo.orangehrmlive.com/web/index.php/auth/login");
-		assertTrue(true);
-	}
-}
